requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

SQLAlchemy==2.0.34
psycopg2-binary==2.9.9

celery==5.4.0
redis==5.0.8

python-multipart==0.0.9

python-docx==1.1.2
pypdf==5.0.1
docx2txt==0.8

structlog==24.4.0
jinja2==3.1.4